App_streamlit/requirements.txt
Keras==2.4.3
numpy==1.22.3
Pillow==9.1.1
seaborn==0.11.2
streamlit==1.10.0
tensorflow==2.3.0
